Gotta move this to analytics package for upcoming stuff
